frontend/requirements.txt
flask
bcrypt
google-cloud-pubsub

#EOF
